Split headfull and headless tests
